+++ app/blog-toaster.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { Transition } from "@headlessui/react";
 import { resolveValue, Toaster } from "react-hot-toast";
-import { CircleCheckIcon } from "./shared/icons";
+import { CircleCheckIcon } from "../common/icons";
 
 export function BlogToaster() {
   return (
