+++ common/util.ts
@@ -1,4 +1,5 @@
 import { Session } from "next-auth";
+import { BlogSession } from "./types";
 
 const siteTitle = "Felix Mokross";
 export const alternateSiteTitle = "felixmokross.dev";
@@ -16,6 +17,6 @@ export function getTitle(pageTitle?: string) {
   return `${pageTitle} \u00B7 ${siteTitle}`;
 }
 
-export function isAuthorized(session: Session) {
+export function isAuthorized(session: BlogSession) {
   return session.login === process.env.NEXT_PUBLIC_GITHUB_USERNAME;
 }
