+++ tailwind.config.js
@@ -4,7 +4,7 @@ const forms = require("@tailwindcss/forms");
 const iOSHeight = require("@rvxlab/tailwind-plugin-ios-full-height");
 
 module.exports = {
-  content: ["./app/**/*.{js,ts,jsx,tsx}", "./src/**/*.{js,ts,jsx,tsx}"],
+  content: ["./app/**/*.{js,ts,jsx,tsx}", "./common/**/*.{js,ts,jsx,tsx}"],
   theme: {
     extend: {
       fontFamily: {
